requirements.txt
torch
transformers
opencv-python
pillow
sentencepiece